Add support for parsers that take NSData
A parser may only accept NSData as input. Therefore add the method
objectForData:error: to the RKParser protocol. If the parsers responds
to that method, it will be invoked with the raw data; if not, the data
will be converted to a NSString and objectForString:error: is invoked.

This will make the implementation of binary formats such as BSON and
MessagePack as parsers easier and faster because the data is not
converted back and forth between NSData and NSString.

Furhter, this commit DRYes up the code that is found in multiple places
to get a parser for a certain MIME type, verify if it actually exists,
and parse the data.