Add support for parser that serialize to NSData
A parser may only serialize to NSData. Therefore add the method
dataForObject:error: to the RKParser protocol. If the parser responds
to that method, it will be invoked; if not, stringForObject:error: will
be invoked and the string is subsequently converted to NSData.

This will make the implementation of binary formats such as BSON and
MessagePack as parsers easier and faster because the data is not
converted back and forth between NSData and NSString.

Furhter, this commit DRYes up the code that is found in multiple places
to get a parser for a certain MIME type, verify if it actually exists,
and serialize the data.