Syncing API is non-blocking.
Adds an RKRequestQueue for now, in the future this should maybe be a different class so that we can group actions together; i.e. a lot of different push requests could be sent at the same time, but a pull should wait until they're finished. A priority queue might be good here.