Introduce the `RKValueTransformer` API providing extensible value transformation to the Object Mapping layer of RestKit. closes #294
This changeset introduces a new API that enables the complete customization of the RestKit object mapping process. All previously
internal value transformations have been transitioned into composable value transformer objects. Value transformers are defined via
a simple Objective-C protocol called `RKValueTransforming` and the library includes a concrete block-based implementation of the protocol
via the `RKBlockValueTransformer` class. Transformers are composable and can be assembled into a transformation chain with precedence based
on index order via the `RKCompoundValueTransformer` class.

Transformers are integrated with the `RKObjectMapping` and `RKPropertyMapping` classes via a single `valueTransformer` property that specifies
what transformer the object mapper should use while mapping a given value.

The core functionality of `RKValueTransformer` is fully reusable and will be made available as a standalone library in the near future.