Fix request descriptor lookup class matching.
`Class` matching was done based on address equality of the `Class`.
Almost always that is enough, but two different `Class` structs can be
created to represent the same thing.  When this happens RestKit could
not match the registered `RequestDescriptor`.

This happened to me and I have no way to create a good reproducable
test.  I found a [StackOverflow
thread](http://stackoverflow.com/questions/16424298/why-is-class-nsclass
fromstringnsstringfromclass-class-on-os-x) on this happening.