incorrect context
If the function is passed a temporary managed object, it will return
nil, instead of the original object not on the destination queue.
Because -[RKRefetchingMappingResult refetchedMappingResult] would create
a copy of the mapped results and overwrite their entries with the
objects in the correct context, it also needed to be updated, since nil
results wouldn't overwrite the original value.