Changed the `RKStringTokenizer` method `- (NSSet *)tokenize:(NSString *)string` to use `kCFStringTokenizerUnitWordBoundary` instead of `kCFStringTokenizerUnitWord`. This then allows tokens to be created for symbols such as '@', '$' and '#' and therefore making them searchable.
As a result of using `kCFStringTokenizerUnitWordBoundary` instead of `kCFStringTokenizerUnitWord`, spaces are detected and a token created. As we don't want to be returning results that include a space, this token (if it exists) is removed before the tokens are returned.

Unit tests added to cover both the creating of tokens from strings that contain symbols and searching for managed objects that contain symbols.